poistettu backend, dockercompose, sql ja tästä vois olla kivampi jatkaa :D
