+++ .gitignore
@@ -29,4 +29,4 @@ dist-ssr
 # expo dir
 .expo/
 
-cmd-helper.ipynb
+run-all.bat
