--- backend/Dockerfile
@@ -1,12 +1,0 @@
-FROM node:16
-
-WORKDIR /app
-
-COPY package*.json ./
-RUN npm install
-
-COPY . .
-
-EXPOSE 8080
-
-CMD ["node", "index.js"]
